+++ CHANGELOG.markdown
@@ -1,6 +1,8 @@
 2.5
 ---
-* Added `Data.Plated` -- a port of Neil Mitchell's `uniplate` that can be used on a `Traversal`.
+* Added `Control.Plated` -- a port of Neil Mitchell's `uniplate` that can be used on a `Traversal`.
+* Added `Data.Data.Lens`
+* Added `Data.Typeable.Lens` with `traverseCast` like `traverseData`
 * Renamed `IndexedStore` to `Context` now that it is used in user-visible locations, and since I also use it as `uniplate`'s notion of a context.
 * Added `Comonad` instances for `Context` and `Kleene`, so we can use stores as the notion of an editable context in uniplate
 * Compatibility with both sets of template haskell quirks for GHC 7.6.1-rc1 and the GHC 7.6.1 development head.